requirements.txt
requests
beautifulsoup4
pillow
torch
transformers
supabase
python-dotenv
